+++ Sources/Adwaita/GtkWidgets/Stack.swift
@@ -61,8 +61,8 @@ public final class Stack: Widget {
     }
 
     /// The transition duration in milliseconds.
-    public var transitionDuration: UInt32 {
-        get { gtk_stack_get_transition_duration(opaquePointer) }
-        set { gtk_stack_set_transition_duration(opaquePointer, newValue) }
+    public var transitionDuration: Int {
+        get { Int(gtk_stack_get_transition_duration(opaquePointer)) }
+        set { gtk_stack_set_transition_duration(opaquePointer, UInt32(newValue)) }
     }
 }
